refactoring for migration support
