+++ nzshm_model/psha_adapter/openquake/logic_tree.py
@@ -31,7 +31,7 @@ from typing import TYPE_CHECKING, Any, Iterator, List, Union
 from lxml import objectify
 
 if TYPE_CHECKING:
-    from nzshm_model.source_logic_tree import logic_tree as slt
+    from nzshm_model.source_logic_tree.version1 import logic_tree as slt
 
 from .uncertainty_models import (
     GenericUncertaintyModel,
