+++ nzshm_model/psha_adapter/openquake/uncertainty_models.py
@@ -8,7 +8,7 @@ from typing import TYPE_CHECKING, Iterator, List
 from lxml import objectify
 
 if TYPE_CHECKING:
-    from nzshm_model.source_logic_tree import logic_tree as slt
+    from nzshm_model.source_logic_tree.version1 import logic_tree as slt
 
     from .logic_tree import LogicTreeBranch
