+++ nzshm_model/source_logic_tree/core/__init__.py
@@ -1,0 +1,1 @@
+from .fault_system_logic_tree import FaultSystemLogicTreeBase, FaultSystemLogicTreeSpec
