+++ nzshm_model/source_logic_tree/version1/logic_tree.py
@@ -11,13 +11,14 @@ from functools import reduce
 from itertools import product
 from math import isclose
 from operator import add, mul
-from typing import Dict, Generator, Iterable, List, Type, Union
+from typing import Dict, Iterable, List, Type, Union
 
 import dacite
 
 from nzshm_model.psha_adapter import PshaAdapterInterface
 
-from .. import BranchAttributeSpec, BranchAttributeValue
+from .. import BranchAttributeValue
+from ..core import FaultSystemLogicTreeBase, FaultSystemLogicTreeSpec
 
 
 @dataclass
@@ -32,62 +33,8 @@ class Branch:
 
 
 @dataclass
-class FaultSystemLogicTreeSpec:
-    short_name: str
-    long_name: str
-    branches: List['BranchAttributeValue'] = field(default_factory=list)
-
-
-@dataclass
-class FaultSystemLogicTree:
-    short_name: str
-    long_name: str
-    branches: List['Branch'] = field(default_factory=list)
-
-    def validate_weights(self) -> bool:
-        weight = 0.0
-        for b in self.branches:
-            weight += b.weight
-        return weight == 1.0
-
-    def derive_spec(self) -> FaultSystemLogicTreeSpec:
-        fslt_spec = FaultSystemLogicTreeSpec(short_name=self.short_name, long_name=self.long_name)
-
-        options: Dict[str, set] = {}
-
-        # extract unique values in to options
-        for branch in self.branches:
-            # iterate all the branches, yielding the unique  branch options
-            # print('branch', branch)
-            for value in branch.values:
-                if value.name not in options.keys():
-                    options[value.name] = set([])
-                val = value.value
-                if isinstance(val, list):  # make it hashable
-                    val = tuple(val)
-                bao = BranchAttributeValue(value.name, value.long_name, val)
-                options[value.name].add(bao)
-
-        # print(options)
-
-        def option_values(options: Dict) -> Generator:
-            # boil down the options values
-            ret = {}
-            for key, opt in options.items():
-                if key not in ret:
-                    ret[key] = BranchAttributeSpec(key, list(opt)[0].long_name)
-
-                values = []
-                for val in opt:
-                    values.append(val.value)
-                ret[key].value_options = sorted(values)
-
-            for key, bas in ret.items():
-                yield bas
-
-        fslt_spec.branches = list(option_values(options))
-
-        return fslt_spec
+class FaultSystemLogicTree(FaultSystemLogicTreeBase):
+    branches: List[Branch] = field(default_factory=list)
 
 
 @dataclass
@@ -137,7 +84,7 @@ class SourceLogicTree:
         ltv = data.get("logic_tree_version")
         if ltv:
             raise ValueError(f"supplied json `logic_tree_version={ltv}` is not supported.")
-        return dacite.from_dict(data_class=SourceLogicTree, data=data)
+        return dacite.from_dict(data_class=SourceLogicTree, data=data, config=dacite.Config(strict=True))
 
     @staticmethod
     def from_json(json_path: Union[pathlib.Path, str]):
