+++ nzshm_model/source_logic_tree/version2/__init__.py
@@ -1,1 +1,1 @@
-from .logic_tree import SourceLogicTree
+from .logic_tree import Branch, SourceLogicTree
